requirements.txt
streamlit
psycopg2-binary
streamlit_authenticator